requirements.txt
async-timeout==5.0.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
gevent==25.8.2
gevent-websocket==0.10.1
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httplib2==0.30.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lrc-kit==0.1.8.3
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.4.0
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2